arkad-sql-demo/sql-market-agent/pyproject.toml
[tool.poetry]
name = "sql-market-agent"
version = "0.1.0"
description = ""
authors = ["olegDk <oleg.davydiuk1995@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = {version = ">=0.1.0"}
openai = "^1.3.7"
black = "^23.11.0"
tavily-python = "^0.1.9"
python-dotenv = "^1.0.0"
faker = "^20.1.0"
python-dateutil = "^2.8.2"
psycopg2-binary = "^2.9.9"
langchain-experimental = "^0.0.49"
langchain-openai = "^0.0.3"
pandas = "^2.2.0"
yfinance = "^0.2.36"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
